Make it easier to manipulate the UI via javascript.
The popup now includes all elements on page load and
only the notification list contents are fetched from
the server via AJAX. This way the elements are always
available for javascript to manipulate.

All the lists and elements now also have an id or a
class that can be used to select them from javascript.